update the parse_async_safely.test.ts and ampx.ts
